+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Common & useful classes, resources, extensions. Based on Symfony framework.
 20. Base Dependency Injection (DI) extension, the `BaseExtension` class
 21. Add main class of this bundle, the `MeritooCommonBundle` class
 22. Add Dependency Injection (DI) Extension (and configuration) for this bundle
+23. Tests > add kernel
 
 # 0.0.1